Поправлены ответы ручек

+++ src/main/java/ru/korett/biometriclab/parameter/ParameterDTO.java
@@ -17,5 +17,5 @@ public class ParameterDTO {
     private String name;
     private String description;
     private String formula;
-    private Integer accuracyPercentage;
+    private Float accuracyPercentage;
 }
